refactor: move ProcessRunner and ProcessRunnerTest to services package, refs #23
